(#41) fix: responsive design (2)
